+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to the "cursorCharCode" extension will be documented in this file.
 
+### v0.9.1 (2020/5/21)
+
+- Rename the extension.
+- Unicode character input helper. (just try Ctrl+I)
+- Fix UTF-8 bugs.
+- Support Python UTF-32 style in hover hint mode.
+
 ### v0.9.0 (2020/5/18)
 
 - Beta release.